# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini ERP API Doc",
        "description": "Documentation for the Mini ERP API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/stocks": {
            "get": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Lister tous les produits en stock",
                "description": "Display a listing of the resource.",
                "operationId": "bc8ad6c6c8cfc92dfda21cb43d84b81f",
                "responses": {
                    "200": {
                        "description": "Liste des produits en stock"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Ajouter un nouveau produit au stock",
                "operationId": "9dd340d8b0489a1d9a50696bda7bcc2f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStock"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé avec succès"
                    },
                    "400": {
                        "description": "Requête invalide"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/stocks/{id}": {
            "get": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Afficher les détails d'un produit spécifique en stock",
                "operationId": "8e09308a73aa8aa31dc6b00f62291a56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du produit"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Mettre à jour les informations d'un produit en stock",
                "operationId": "8d4257795eb700aae904a7de5cf941f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit mis à jour avec succès"
                    },
                    "400": {
                        "description": "Requête invalide"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "ID du produit",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom du produit",
                        "type": "string"
                    },
                    "sku": {
                        "description": "SKU du produit",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description du produit",
                        "type": "string"
                    },
                    "is_active": {
                        "description": "Statut actif du produit",
                        "type": "boolean"
                    },
                    "price": {
                        "description": "Prix du produit",
                        "type": "number",
                        "format": "float"
                    },
                    "quantity": {
                        "description": "Quantité en stock",
                        "type": "integer"
                    },
                    "low_stock_threshold": {
                        "description": "Seuil de stock bas",
                        "type": "integer"
                    },
                    "category_id": {
                        "description": "ID de la catégorie",
                        "type": "integer"
                    },
                    "attributes": {
                        "description": "Attributs supplémentaires du produit",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ProductStock": {
                "required": [
                    "product_id",
                    "quantity",
                    "location_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "location_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}